+++ src/main/java/backend/verify.java
@@ -1,6 +1,11 @@
 package backend;
 
+import javax.swing.*;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
 import java.util.Objects;
 
 /**
@@ -42,6 +47,47 @@ public class verify {
     public boolean verify_string(String string, String hash) {
         String new_hash = hashing.string(string);
         return compare_hashes(hash, new_hash);
+    }
+
+    /**
+     * This checks the checksum file and verifies the integrity of the file
+     * @param path_to_file The path to the file you want to verify
+     * @param hashText The main status bar
+     */
+    public void verify_with_checksum(String path_to_file, JTextField hashText) {
+        try {
+
+            // Create a class for the common functions
+            common Common = new common();
+
+            // Make 'Path' objects
+            Path dir = Paths.get(path_to_file).getParent();
+            Path shafile = Paths.get(dir.toString(), "sha256.txt");
+
+            // Get the first line of the file
+            List<String> contents = Files.readAllLines(shafile);
+            String line = contents.get(0);
+
+            // Split the line and obtain information
+            String[] line_split = line.split("  ");
+            String hash = line_split[0];
+            String filename = line_split[1];
+
+            // Verifies the file if the filename is in the absolute path
+            boolean samefile = Common.filename_in_path(path_to_file, filename);
+            if (samefile) {
+                if (verify_file(path_to_file, hash)) {
+                    hashText.setText("File Is The Same!");
+                } else {
+                    hashText.setText("File Is Not The Same!");
+                }
+            } else {
+                hashText.setText("This Is Not The File That's Hashed!");
+            }
+
+        } catch (IOException ex) {
+            hashText.setText("There is no 'sha256.txt' in this DIR!");
+        }
     }
 
 }