+++ CHANGELOG.md
@@ -47,8 +47,17 @@ This release added JavaDoc / API Docs for all the classes and their functions
 ### Fixed
 - None
 
+## [v1.1.3] - 13th December, 2021
+### Notes
+This release was a code cleaning and maintenance release 
+### Features
+- Made the verify class useful
+### Fixed
+- Issue of the filename not being detected in the absolute path
+
 [v1.0]: https://github.com/Smilin-Dominator/Devify/releases/tag/v1.0
 [v1.0.1]: https://github.com/Smilin-Dominator/Devify/compare/v1.0...v1.0.1
 [v1.1]: https://github.com/Smilin-Dominator/Devify/compare/v1.0.1...v1.1
 [v1.1.1]: https://github.com/Smilin-Dominator/Devify/compare/v1.1...v1.1.1
 [v1.1.2]: https://github.com/Smilin-Dominator/Devify/compare/v1.1.1...v1.1.2
+[v1.1.3]: https://github.com/Smilin-Dominator/Devify/compare/v1.1.2...v1.1.3